refactor: address CodeRabbit review suggestions
- Update build_trash_check / build_lateral_trash_exclusion doc comments
  to match behavior: returns TRUE (no exclusion) when the link has no
  TRASH label, since a missing TRASH label means no message can be
  trashed. The doc previously promised a name+link_id fallback that
  never existed.
- Add #[tracing::instrument(skip(pool, ast), err)] to resolve_filters
  per project guideline that fns returning Result use tracing::instrument
  with err.
- Move fold_unresolved / can_short_circuit unit tests out of the inline
  `mod fold_tests` block in resolve.rs into dynamic/tests.rs, matching
  the project convention of putting tests in a sibling tests.rs rather
  than inline cfg(test) blocks. fold_unresolved and collect_complete_emails
  bumped to pub(super) so the tests can reach them.
- Drop confusing double-negative assert in
  test_combined_thread_id_and_sender_splits_correctly — line 403 already
  asserts the SQL contains "from_contact_id"; the redundant
  !has_no_raw_containing was hard to read.
